# Add TEST_SKIP_GUEST_BUILD env to skip re-building guests