@eyusufatik 2025-03-21
wanna update the make test command?